[수정] 감정값 db에서 가져와서 추천main 연결 / 장르도 가져와서 추천

+++ src/main/java/com/example/musicemotion/dto/EmotionDTO.java
@@ -4,16 +4,16 @@ public class EmotionDTO {
     private int emotion_id;
     private String emotion_name;
     private String description;
-    private double danceability_min;
-    private double danceability_max;
-    private double energy_min;
-    private double energy_max;
-    private double loudness_min;
-    private double loudness_max;
-    private double tempo_min;
-    private double tempo_max;
-    private double valance_min;
-    private double valance_max;
+    private float danceability_min;
+    private float danceability_max;
+    private float energy_min;
+    private float energy_max;
+    private float loudness_min;
+    private float loudness_max;
+    private float tempo_min;
+    private float tempo_max;
+    private float valance_min;
+    private float valance_max;
 
     // Getters and Setters
     public int getEmotion_id() {
@@ -34,64 +34,64 @@ public class EmotionDTO {
     public void setDescription(String description) {
         this.description = description;
     }
-    public double getDanceability_min() {
+    public float getDanceability_min() {
         return danceability_min;
     }
-    public void setDanceability_min(double danceability_min) {
+    public void setDanceability_min(float danceability_min) {
         this.danceability_min = danceability_min;
     }
-    public double getDanceability_max() {
+    public float getDanceability_max() {
         return danceability_max;
     }
-    public void setDanceability_max(double danceability_max) {
+    public void setDanceability_max(float danceability_max) {
         this.danceability_max = danceability_max;
     }
-    public double getEnergy_min() {
+    public float getEnergy_min() {
         return energy_min;
     }
-    public void setEnergy_min(double energy_min) {
+    public void setEnergy_min(float energy_min) {
         this.energy_min = energy_min;
     }
-    public double getEnergy_max() {
+    public float getEnergy_max() {
         return energy_max;
     }
-    public void setEnergy_max(double energy_max) {
+    public void setEnergy_max(float energy_max) {
         this.energy_max = energy_max;
     }
-    public double getLoudness_min() {
+    public float getLoudness_min() {
         return loudness_min;
     }
-    public void setLoudness_min(double loudness_min) {
+    public void setLoudness_min(float loudness_min) {
         this.loudness_min = loudness_min;
     }
-    public double getLoudness_max() {
+    public float getLoudness_max() {
         return loudness_max;
     }
-    public void setLoudness_max(double loudness_max) {
+    public void setLoudness_max(float loudness_max) {
         this.loudness_max = loudness_max;
     }
-    public double getTempo_min() {
+    public float getTempo_min() {
         return tempo_min;
     }
-    public void setTempo_min(double tempo_min) {
+    public void setTempo_min(float tempo_min) {
         this.tempo_min = tempo_min;
     }
-    public double getTempo_max() {
+    public float getTempo_max() {
         return tempo_max;
     }
-    public void setTempo_max(double tempo_max) {
+    public void setTempo_max(float tempo_max) {
         this.tempo_max = tempo_max;
     }
-    public double getValance_min() {
+    public float getValance_min() {
         return valance_min;
     }
-    public void setValance_min(double valance_min) {
+    public void setValance_min(float valance_min) {
         this.valance_min = valance_min;
     }
-    public double getValance_max() {
+    public float getValance_max() {
         return valance_max;
     }
-    public void setValance_max(double valance_max) {
+    public void setValance_max(float valance_max) {
         this.valance_max = valance_max;
     }
 }
